Verify 0.11.0 for the first time, and fix the harness that prevented it
0.11.0 shipped source-only and said so: the session that wrote it had no working
Unity process, so the LOD ladder and the SOG scaffolding had never been
compiled, let alone run. This session has Unity, so the obvious move was to run
the verifier — which then spun for 22 minutes without finishing.

The harness had a non-terminating loop. `for (byte tag = 252; tag <= 255;
tag++)` cannot end, because a byte never exceeds 255: the counter wraps to 0 and
goes round again. It logged 12.5 million failing iterations and grew verify.log
to 11.4 GB, still climbing at ~5 MB/s, and never printed a result. The first
four iterations passed and looked exactly like a working test, which is the
whole problem with it — what followed read as "the quaternion port is broken"
and was nothing of the kind. The counter is an int now, cast at the call.

With that fixed the run finishes and 0.11.0 is verified: zero compile errors on
Unity 6000.3.22f1 + URP 17.3.0, 24 checks all pass. The LOD symbols are in the
assembly; SogFileReader and the ISogTextureDecoder seam compile; DequantizeQuat
returns a unit quaternion on all four maxComp branches, which is the first
actual execution of the port taken from @playcanvas/splat-transform's source.

Running a real Editor against the package also generated the .meta file 0.11.0
recorded as deliberately missing — it was waiting for exactly this, rather than
risking a hand-written GUID. Checked unique before committing.

Neither feature is finished, and the docs now say which half is done: SOG
compiles but still cannot decode WebP, so no .sog has been read end to end; the
LOD ladder compiles but has never run against a capture.

Also: the "17 commits ahead" figure in three documents was stale the moment it
was written and rots on every commit. Replaced with the fact that actually
matters, 0 behind, with the ahead-count given as a dated observation to
re-count rather than quote.

Adds AGENTS.md, a briefing for an agent picking up work here.

+++ tools/verify/PackageVerify.cs
@@ -127,11 +127,20 @@ public static class PackageVerify
             var dequantizeQuat = sogType.GetMethod("DequantizeQuat", BindingFlags.Public | BindingFlags.Static);
             if (dequantizeQuat != null)
             {
-                for (byte tag = 252; tag <= 255; tag++)
+                // The loop counter is an INT, cast to byte at the call. Written
+                // as `for (byte tag = 252; tag <= 255; tag++)` it never
+                // terminates: a byte cannot exceed 255, so the condition is
+                // always true and the counter wraps to 0 and goes round again.
+                // Measured 2026-09-10 before the fix -- the run spun for 22
+                // minutes, logged 12.5 million failing iterations and grew
+                // verify.log to 11.4 GB, still climbing at ~5 MB/s, with no
+                // VERIFY line ever printed. A harness that cannot finish is
+                // worse than one that fails.
+                for (int tag = 252; tag <= 255; tag++)
                 {
                     try
                     {
-                        var q = (UnityEngine.Quaternion)dequantizeQuat.Invoke(null, new object[] { (byte)37, (byte)200, (byte)90, tag });
+                        var q = (UnityEngine.Quaternion)dequantizeQuat.Invoke(null, new object[] { (byte)37, (byte)200, (byte)90, (byte)tag });
                         float mag = Mathf.Sqrt(q.x * q.x + q.y * q.y + q.z * q.z + q.w * q.w);
                         Check($"DequantizeQuat(tag={tag}) returns a unit quaternion",
                               Mathf.Abs(mag - 1f) < 0.001f,
